+++ ide/src/rust/ide/src/controller/graph/executed.rs
@@ -292,6 +292,12 @@ impl Context for Handle {
     }
 }
 
+impl model::undo_redo::Aware for Handle {
+    fn undo_redo_repository(&self) -> Rc<model::undo_redo::Repository> {
+        self.graph.borrow().undo_redo_repository()
+    }
+}
+
 
 
 // ============
@@ -325,8 +331,10 @@ pub mod tests {
 
     impl MockData {
         pub fn controller(&self) -> Handle {
+            let logger      = Logger::new("test");
             let parser      = parser::Parser::new_or_panic();
-            let module      = self.module.plain(&parser);
+            let repository  = Rc::new(model::undo_redo::Repository::new(&logger));
+            let module      = self.module.plain(&parser,repository);
             let method      = self.graph.method();
             let mut project = model::project::MockAPI::new();
             let ctx         = Rc::new(self.ctx.create());
@@ -340,7 +348,7 @@ pub mod tests {
             let suggestion_db = self.graph.suggestion_db();
             model::project::test::expect_suggestion_db(&mut project,suggestion_db);
             let project = Rc::new(project);
-            Handle::new(Logger::new("test"),project.clone_ref(),method).boxed_local().expect_ok()
+            Handle::new(logger,project.clone_ref(),method).boxed_local().expect_ok()
         }
     }