+++ ide/src/rust/ide/src/model.rs
@@ -21,9 +21,10 @@
 
 pub mod execution_context;
 pub mod module;
+pub mod project;
 pub mod registry;
 pub mod suggestion_database;
-pub mod project;
+pub mod undo_redo;
 
 pub use execution_context::ExecutionContext;
 pub use module::Module;
@@ -37,4 +38,5 @@ pub mod traits {
     pub use execution_context::API as TRAIT_ExecutionContextApi;
     pub use module::API            as TRAIT_ModuleApi;
     pub use project::API           as TRAIT_ProjectApi;
+    pub use undo_redo::Aware       as TRAIT_UndoRedoAware;
 }